Moved ConsoleButtons to layouts added All. renamed files to be .jsx Deleted unused css in App.css
